# api/openapi.yaml
# Design-time API contract for this service domain — owned BY this repo
# (contract-per-repo, Netflix-style; no central contracts repo).
# The runtime spec served at /v3/api-docs must stay compatible with this file;
# Phase 2 adds contract tests that enforce it.
openapi: 3.0.3
info:
  title: "Shariah Audit — BIAN Semantic API"
  description: >
    BIAN Service Domain 'Shariah Audit' (functional pattern: Assess,
    business area: Shariah Governance and Compliance) operating on the
    'Shariah Audit Engagement Assessment' control record.
  version: 0.1.0
servers:
  - url: /
    description: direct (in-cluster)
  - url: /isb-shariah-audit
    description: through the platform gateway (prefix is rewritten away)
tags:
  - name: service-domain
  - name: control-record

paths:
  /v1/service-domain:
    get:
      tags: [service-domain]
      operationId: getServiceDomain
      summary: Service domain metadata
      responses:
        "200":
          description: SD identity and BIAN classification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDomainInfo"

  /v1/shariah-audit-engagement-assessment/initiate:
    post:
      tags: [control-record]
      operationId: initiate
      summary: "BIAN action term: Initiate"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Properties"
      responses:
        "201":
          description: Control record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlRecord"

  /v1/shariah-audit-engagement-assessment:
    get:
      tags: [control-record]
      operationId: list
      summary: "BIAN action term: Retrieve (collection)"
      responses:
        "200":
          description: All control records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ControlRecord"

  /v1/shariah-audit-engagement-assessment/{crId}/retrieve:
    get:
      tags: [control-record]
      operationId: retrieve
      summary: "BIAN action term: Retrieve"
      parameters:
        - $ref: "#/components/parameters/CrId"
      responses:
        "200":
          description: The control record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlRecord"
        "404":
          description: Unknown control record id

  /v1/shariah-audit-engagement-assessment/{crId}/update:
    put:
      tags: [control-record]
      operationId: update
      summary: "BIAN action term: Update"
      parameters:
        - $ref: "#/components/parameters/CrId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Properties"
      responses:
        "200":
          description: Updated control record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlRecord"
        "404":
          description: Unknown control record id

  /v1/shariah-audit-engagement-assessment/{crId}/control:
    put:
      tags: [control-record]
      operationId: control
      summary: "BIAN action term: Control (suspend | resume | terminate)"
      parameters:
        - $ref: "#/components/parameters/CrId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlAction"
      responses:
        "200":
          description: Control applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlRecord"
        "400":
          description: Unknown control action
        "404":
          description: Unknown control record id

components:
  parameters:
    CrId:
      name: crId
      in: path
      required: true
      schema:
        type: string
      description: Control record identifier (CR-…)

  schemas:
    ServiceDomainInfo:
      type: object
      properties:
        serviceDomain: { type: string, example: "Shariah Audit" }
        businessArea: { type: string, example: "Shariah Governance and Compliance" }
        businessDomain: { type: string, example: "Shariah Governance" }
        functionalPattern: { type: string, example: "Assess" }
        assetType: { type: string, example: "Shariah Audit Engagement" }
        controlRecord: { type: string, example: "Shariah Audit Engagement Assessment" }
        version: { type: string }
        phase: { type: string }

    ControlRecord:
      type: object
      properties:
        controlRecordId: { type: string, example: "CR-7f3a…" }
        controlRecordType: { type: string, example: "Shariah Audit Engagement Assessment" }
        status:
          type: string
          enum: [INITIATED, ACTIVE, SUSPENDED, TERMINATED]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        properties:
          $ref: "#/components/schemas/Properties"

    Properties:
      type: object
      additionalProperties: true
      description: Free-form domain properties (Phase 2 replaces with the real per-domain schema)

    ControlAction:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [suspend, resume, terminate]